Removing unused imports

+++ back-office/server/src/main/java/proj/cloud/ath/repositories/GlobalStatRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import proj.cloud.ath.entities.stat.GlobalStat;
 
-public interface GlobalStatRepository extends JpaRepository<GlobalStat, Integer> {
+public interface GlobalStatRepository extends JpaRepository<GlobalStat, Double> {
 
 }
